packages/toolkit/core-kit/src/utils/url.ts
import { looseNativeUriSchemeProtocolRegEx, webRedirectUriProtocolRegEx } from '../regex.js';

export const validateRedirectUrl = (url: string, type: 'web' | 'native') => {
  try {
    const { protocol } = new URL(url);
    const protocolRegEx =
      type === 'native' ? looseNativeUriSchemeProtocolRegEx : webRedirectUriProtocolRegEx;

    return protocolRegEx.test(protocol);
  } catch {
    return false;
  }
};

export const validateUriOrigin = (url: string) => {
  try {
    return new URL(url).origin === url;
  } catch {
    return false;
  }
};
